pyproject.toml
[project]
name = "disting-nt-mcp"
version = "0.1.0"
description = "MCP server for controlling Expert Sleepers Disting NT multi-algorithm DSP module"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "python-rtmidi>=1.5.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
disting-nt-mcp = "server:main"
